pkg: Ignore tags that are incomplete semantic versions
If a package is tagged with an incomplete semantic version (e.g., `v1`), the
package manager will happily parse it and consider it as a candidate for
installation as a dependency (`v1` parses to `v1.0.0`). This can lead to the
wrong version being installed if the incomplete tag points to a different
version and the package has an exact version constraint. Avoid this by filtering
out any versions that don't match the tags they were parsed from.

Note that this will break anyone intentionally trying to depend on a `v1`
tag. However, note also that anyone doing this probably wasn't getting the
behavior they wanted before, since `v1.0.0` would also match a constraint of
`v1`.

Fixes #6772

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>