Optimize circuit breaker locking a little
Mostly only take the write lock when we need it. Also makes the GC
(which happens with the write lock) only happen if we'd be updating the
map to add a new target anyway. This keeps the targets list bounded,
though it does mean stale targets could hang around longer.

Signed-off-by: Nic Cope <nicc@rk0n.org>