WIP: Use invalidatable client in the engine
The idea here is to recreate the engine's clients and REST mappers
whenever a CRD is deleted (i.e. whenever an XRD is deleted). This works
around the fact that the client and REST mapper both cache scope by GVK,
and never invalidate that cache. So the goal is you can delete a cluster
scoped XRD, create a namespaced one with the same GVK, and it'll work.

Unfortunately it doesn't.

Signed-off-by: Nic Cope <nicc@rk0n.org>