Create only one Transaction per package generation
The package controller creates transactions when a package is created.
We only want to do this for 'top-level' packages; those not installed as
dependences of other packages. Packages that're dependencies are handled
by the top-level package's transaction.

At the same time, if a user later edits a package originally installed
as a dependency, we do want to create a new Transaction. We handle this
by recording the last generation we created a Transaction for, and only
creating a Transaction if the generation changes.

Signed-off-by: Nic Cope <nicc@rk0n.org>