Remove version parameter from Installer interface
The Installer.Install method no longer requires a separate version
parameter since xpkg.Package now includes the Version field. All
implementations now use xp.Version instead of receiving version as a
separate argument.

This simplifies the interface and reduces parameter coupling between the
xpkg.Client and the transaction installer. The version is now a property
of the Package itself, which is more intuitive and makes the code easier
to follow.

Signed-off-by: Nic Cope <nicc@rk0n.org>