Retry on NotFound/Conflict when installing
We have a pattern where we write to the API server in one function, then
read it back in another. These are cache backed reads and often the
cache is a little slow. Instead of failing the whole reconcile (and
re-running the whole thing with exponential backoff), just retry failed
reads/writes a few times.

Signed-off-by: Nic Cope <nicc@rk0n.org>