HACK: Restore GVK after calling client.Create in Establisher
For some truly cursed reason that I cannot explain, this call is
removing the GVK from the supplied object. This results in Establisher
returning an array of resourceRefs without apiVersion and kind, which in
turn means you can't release the objects (because that uses the array of
resourceRefs in a revision's status.objects).

Signed-off-by: Nic Cope <nicc@rk0n.org>