pkg: Allow configuration of DeploymentRuntimeConfig via ImageConfig
When installing a package directly, a `DeploymentRuntimeConfig` can be selected
by setting a reference in the package spec. However, there's no way to select a
`DeploymentRuntimeConfig` for packages installed as dependencies, leading to the
common practice if disabling dependency resolution and installing dependencies
manually.

Add a new `runtime` field to the `ImageConfig` allowing a
`DeploymentRuntimeConfig` to be specified for all packages matching a pattern,
regardless of how they're installed. If both are present, the `ImageConfig`
takes precedence over the `runtimeConfigRef` in the package spec.

Fixes #6382

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>