xpkg: Enforce that functions can install only CRDs
The xpkg spec states that functions may include zero or more CRDs and no other
resource types. The linter was not enforcing this, allowing functions to install
other resource types such as XRDs and Compositions.

Add the `IsCRD` validator to the function linter to prevent other types from
being accepted in functions.

Fixes #6975

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>