pkg: Separate validation and linting of packages
Crossplane uses package linters in two places: at build time (in the CLI) and at
install time (in the revision controller). In both cases, linter errors are
treated as errors and stop the operation.

At build time, this makes sense: we should be strict about the packages we
produce. At install time, this means we may reject packages that we could
successfully install. For example (as described in #6525) if we allow new
resource types in a package in a new version of Crossplane (as we've done with
Operations and MRDs), older versions should still be able to install packages
that contain them; the new resources would just be ignored. If package authors
want to limit the versions their package works with, they can set a version
constraint in the metadata.

Introduce a new Validator type, which is identical to a linter except in
name. Add validators for all package types, which are identical to their linters
except that they allow any object to be included.

In the revision controller, use validators to determine whether package
installation should be attempted, where we previously used linters. Run the
linters as well and record any lint errors as events, to help the user diagnose
any unexpected behavior.

Fixes #6525

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>