pkg: Stop installing CRDs from Function packages
The xpkg spec says that CRDs included in Function packages will not be installed
into the cluster, but until now they have been since Functions are treated the
same as all other packages. Additionally, the previous commit allowed packages
with arbitrary objects included to be installed, contrary to the spec (and
desired behavior).

Filter objects in the establisher, so that regardless of what kinds of resources
are present in a package we install only the desired ones.

Fixes #5294

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>