pkg: Move DAG logic out of the apis package
We're going to make the apis directory its own go module so that we can depend
on it from crossplane-runtime, and so that other importers don't need to depend
on all of crossplane. To do that, apis can't depend on anything from internal,
which means we need to implement some DAG-related methods outside of apis. This
is a nice change anyway, since the DAG is an internal Crossplane concern that
shouldn't really be exposed through the API.

Implement wrapper types in the pkg controller that implement the necessary DAG
methods, and update consuming code to use them.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>